Fix plaintext/json
