+++ frameworks/CSharp/aspnetcore/src/Platform/Program.cs
@@ -7,7 +7,6 @@ using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
-using Npgsql;
 
 namespace PlatformBenchmarks
 {
